feat: added yup validation
