+++ package.json
@@ -16,6 +16,7 @@
     "@emotion/server": "^11.10.0",
     "@emotion/styled": "^11.10.0",
     "@headlessui/react": "^1.6.6",
+    "@hookform/resolvers": "^2.9.9",
     "@next-auth/prisma-adapter": "^1.0.4",
     "@prisma/client": "^4.3.1",
     "@reduxjs/toolkit": "^1.8.5",
@@ -39,6 +40,7 @@
     "react-redux": "^8.0.2",
     "swr": "^1.3.0",
     "usehooks-ts": "^2.6.0",
+    "yup": "^0.32.11",
     "zod": "^3.19.1"
   },
   "devDependencies": {
